requirements.txt
python-telegram-bot[job-queue]==20.3
openai==0.28
pytz
aiohttp
